requirements.txt
streamlit
mistralai
webvtt-py